Add new common dd change page
