+++ src/js/common/schemaform/definitions/bankAccount.js
@@ -1,21 +1,19 @@
-import _ from 'lodash/fp';
 import commonDefinitions from 'vets-json-schema/dist/definitions.json';
 import { validateRoutingNumber } from '../validation';
 
-export const schema = _.set(
-  'properties.accountType.enumNames',
-  [
-    'Checking',
-    'Savings'
-  ],
-  commonDefinitions.bankAccount
-);
+export const schema = commonDefinitions.bankAccount;
 
 export const uiSchema = {
   'ui:order': ['accountType', 'accountNumber', 'routingNumber'],
   accountType: {
     'ui:title': 'Account type',
-    'ui:widget': 'radio'
+    'ui:widget': 'radio',
+    'ui:options': {
+      labels: {
+        checking: 'Checking',
+        savings: 'Savings'
+      }
+    }
   },
   accountNumber: {
     'ui:title': 'Account number'
