+++ test/edu-benefits/5495/schema/maximal-test.json
@@ -111,6 +111,7 @@
   },
   "directDeposit": {
     "data": {
+      "bankAccountChange": "startUpdate",
       "bankAccount": {
         "accountType": "checking",
         "routingNumber": "211370545",
