+++ test/edu-benefits/pages/directDeposit.unit.spec.jsx
@@ -4,10 +4,11 @@ import { expect } from 'chai';
 import ReactTestUtils from 'react-addons-test-utils';
 
 import { DefinitionTester } from '../../util/schemaform-utils.jsx';
-import formConfig from '../../../src/js/edu-benefits/pages/directDeposit.js';
+import formConfig1995 from '../../../src/js/edu-benefits/1995/config/form';
+import formConfig5495 from '../../../src/js/edu-benefits/5495/config/form';
 
-describe('Edu page directDeposit', () => {
-  const { schema, uiSchema } = formConfig;
+const pageTests = (page) => {
+  const { schema, uiSchema } = page;
   it('should render', () => {
     const form = ReactTestUtils.renderIntoDocument(
       <DefinitionTester
@@ -18,7 +19,41 @@ describe('Edu page directDeposit', () => {
 
     const formDOM = findDOMNode(form);
 
-    expect(formDOM.querySelectorAll('input').length).to.equal(4);
+    expect(formDOM.querySelectorAll('input').length).to.equal(3);
+  });
+  it('should render stop message', () => {
+    const form = ReactTestUtils.renderIntoDocument(
+      <DefinitionTester
+          schema={schema}
+          data={{}}
+          uiSchema={uiSchema}/>
+    );
+
+    const formDOM = findDOMNode(form);
+    ReactTestUtils.Simulate.change(formDOM.querySelector('#root_bankAccountChange_2'), {
+      target: {
+        value: 'stop'
+      }
+    });
+
+    expect(formDOM.textContent).to.contain('The Department of Treasury requires');
+  });
+  it('should render bank account fields', () => {
+    const form = ReactTestUtils.renderIntoDocument(
+      <DefinitionTester
+          schema={schema}
+          data={{}}
+          uiSchema={uiSchema}/>
+    );
+
+    const formDOM = findDOMNode(form);
+    ReactTestUtils.Simulate.change(formDOM.querySelector('#root_bankAccountChange_1'), {
+      target: {
+        value: 'startUpdate'
+      }
+    });
+
+    expect(formDOM.querySelectorAll('input').length).to.equal(7);
   });
   it('should show error on bad routing number', () => {
     const form = ReactTestUtils.renderIntoDocument(
@@ -29,6 +64,11 @@ describe('Edu page directDeposit', () => {
     );
 
     const formDOM = findDOMNode(form);
+    ReactTestUtils.Simulate.change(formDOM.querySelector('#root_bankAccountChange_1'), {
+      target: {
+        value: 'startUpdate'
+      }
+    });
     const routingNumber = formDOM.querySelector('#root_bankAccount_routingNumber');
     ReactTestUtils.Simulate.blur(routingNumber);
 
@@ -42,4 +82,9 @@ describe('Edu page directDeposit', () => {
 
     expect(formDOM.querySelector('.usa-input-error #root_bankAccount_routingNumber')).not.to.be.null;
   });
+};
+
+describe('Edu directDepositChangePage', () => {
+  describe('1995', () => pageTests(formConfig1995.chapters.personalInformation.pages.directDeposit));
+  describe('5495', () => pageTests(formConfig5495.chapters.personalInformation.pages.directDeposit));
 });
